Add polyfill to fix how `cmd` binds on OSX
On OSX, the cmd module binds using the traditional readline system, but OSX uses
libedit, which has a different interface.  This allows for OSX to handle the tab
completion correctly.  (A patch for cpython was submitted a year ago and denied,
wanting a more in depth solution, only to be ignored entirely since).